+++ e2e/tests/aws.rs
@@ -2,7 +2,7 @@
 mod tests {
     use anyhow::Result;
     use e2e::e2e_helpers::start_aws_kms;
-    use fuels::accounts::kms::AwsWallet;
+    use fuels::accounts::kms::{AwsWallet, KmsSigner};
     use fuels::accounts::{Account, ViewOnlyAccount};
     use fuels::prelude::{
         launch_provider_and_get_wallet, AssetId, Contract, LoadConfiguration, TxPolicies,
@@ -16,7 +16,7 @@ mod tests {
 
         let amount = 500000000;
         let key = kms.create_key().await?;
-        let address = key.kms_key.address().clone();
+        let address = key.kms_key.fuel_address().clone();
 
         wallet
             .transfer(&address, amount, AssetId::zeroed(), TxPolicies::default())
@@ -43,7 +43,7 @@ mod tests {
 
         let amount = 500000000;
         let key = kms.create_key().await?;
-        let address = key.kms_key.address().clone();
+        let address = key.kms_key.fuel_address().clone();
 
         wallet
             .transfer(&address, amount, AssetId::zeroed(), TxPolicies::default())
