+++ packages/fuels-accounts/src/kms.rs
@@ -1,5 +1,13 @@
 mod aws;
 mod google;
+mod kms_trait;
+mod signature_utils;
+mod wallet_types;
 
 pub use aws::*;
+pub use aws_config;
+pub use aws_sdk_kms;
 pub use google::*;
+pub use google_cloud_kms;
+pub use kms_trait::*;
+pub use wallet_types::*;
