+++ packages/fuels-accounts/src/kms/aws.rs
@@ -1,5 +1,2 @@
-mod wallet;
-pub use wallet::*;
-
-pub use aws_config;
-pub use aws_sdk_kms;
+mod signer;
+pub use signer::*;
