+++ packages/fuels-accounts/src/kms/google.rs
@@ -1,4 +1,2 @@
-mod wallet;
-pub use wallet::*;
-
-pub use google_cloud_kms;
+mod signer;
+pub use signer::*;
